requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-classic
validators
youtube-transcript-api
unstructured
pytubefix
bs4
lxml
python-dotenv
streamlit